Conflicts:
	PdChocolate/@chocolate.workspace.layout
	PdChocolate/c.bang/c.bang.layout
	PdChocolate/c.blackboard/c.blackboard.layout
	PdChocolate/c.colorpanel/c.colorpanel.layout
	PdChocolate/c.gain_tilde/c.gain~.layout
	PdChocolate/c.incdec/c.incdec.layout
	PdChocolate/c.knob/c.knob.layout
	PdChocolate/c.menu_ui/c.menu.layout
	PdChocolate/c.meter_tilde/c.meter~.layout
	PdChocolate/c.number/c.number.layout
	PdChocolate/c.number_tilde/c.number~.layout
	PdChocolate/c.plane/c.plane.layout
	PdChocolate/c.preset/c.preset.layout
	PdChocolate/c.tab/c.tab.layout

+++ Sources/eclass/eclass_attrgetter.c
@@ -1,7 +1,7 @@
 /*
- * PdEnhanced - Pure Data Enhanced
+ * CicmWrapper - Pure Data Enhanced
  *
- * An add-on for Pure Data
+ * A wrapper for Pure Data
  *
  * Copyright (C) 2013 Pierre Guillot, CICM - Université Paris 8
  * All rights reserved.
@@ -55,9 +55,9 @@ void eclass_attr_getter(t_object* x, t_symbol *s, int* argc, t_atom** argv)
             else
             {
                 point = (char *)x + c->c_attr[i].size;
-                argc[0] = (long)point[0];
+                argc[0] = (int)point[0];
                 if(argc[0] > c->c_attr[i].sizemax)
-                    argc[0] = c->c_attr[i].sizemax;
+                    argc[0] = (int)c->c_attr[i].sizemax;
             }
 
             point = (char *)x + c->c_attr[i].offset;
